Baseline
# Main Class:
Test class. This is the class that will be ran on program start. I made a simple program demonstrating what the program is capable of so far.
# Entity:
Everything alive will be an Entity, including the player. This is an abstract class
# Player:
This is the player character class.
# displayValue:
This class allows to display values. It will mostly be used to display Monster HP, player HP, but it can also be used to display MP, and other values.